Fix additional warnings
(cherry picked from commit b8daf96a618617513122da253a024c1a6909d910)

+++ src/backend/cpu/Array.hpp
@@ -218,7 +218,7 @@ namespace cpu
         int useCount() const
         {
             if (!data.get()) eval();
-            return data.use_count();
+            return static_cast<int>(data.use_count());
         }
 
         operator Param<T>()
